refactor: index 파일명 변경
