+++ src/assets/images/loginImageSlider/loginImagesList.ts
